tu2019: Fix 'fixed' revision to use proper android manifest
We fixed product to commit that was used for 'release' version.
But android manifest moved further, so we need to fix not only product
but also android.
Android manifest is fixed as last commit of product.
So this is the reason that we need to fix product not to commit that
was used for release but for commit with fixed android.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>